Move TextInput component to widget.rs
It is unrelated to query
